# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "企业级在线考试系统，支持考题管理、考试配置、考生管理、成绩统计与导出。\n认证方式：Bearer Token（通过 /api/auth/login 获取，Cookie 或 Authorization Header 传递）",
        "title": "在线考试系统 API",
        "termsOfService": "http://svip.com/terms",
        "contact": {
            "name": "API 支持",
            "email": "admin@svip.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "2.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/api/admin/questions": {
            "get": {
                "security": [
                    {
                        "BearerToken": []
                    }
                ],
                "description": "分页查询题库题目，支持按分类、关键词、题型筛选",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "题目管理"
                ],
                "summary": "获取题目列表",
                "parameters": [
                    {
                        "type": "string",
                        "description": "分类 ID",
                        "name": "categoryId",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "搜索关键词（题干）",
                        "name": "keyword",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "题型：single/multiple/judge/fill/essay",
                        "name": "type",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "页码 (默认 1)",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "每页条数 (默认 10)",
                        "name": "pageSize",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "题目列表",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/models.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/models.PagedData"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "查询失败",
                        "schema": {
                            "$ref": "#/definitions/models.APIResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerToken": []
                    }
                ],
                "description": "新增一道题目到题库",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "题目管理"
                ],
                "summary": "创建题目",
                "parameters": [
                    {
                        "description": "题目信息",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.CreateQuestionReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "添加成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/models.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/models.Question"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "参数错误",
                        "schema": {
                            "$ref": "#/definitions/models.APIResponse"
                        }
                    },
                    "500": {
                        "description": "创建失败",
                        "schema": {
                            "$ref": "#/definitions/models.APIResponse"
                        }
                    }
                }
            }
        },
...
        "/api/admin/questions/import": {
            "post": {
                "security": [
                    {
                        "BearerToken": []
                    }
                ],
                "description": "通过上传文件（json/csv/xlsx）批量导入题目到指定分类",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "题目管理"
                ],
                "summary": "批量导入题目",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "分类 ID",
                        "name": "categoryId",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "file",
                        "description": "导入文件 (.json/.csv/.xlsx)",
                        "name": "file",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "导入成功",
                        "schema": {
                            "$ref": "#/definitions/models.APIResponse"
                        }
                    },
                    "400": {
                        "description": "参数错误或文件格式错误",
                        "schema": {
                            "$ref": "#/definitions/models.APIResponse"
                        }
                    },
                    "500": {
                        "description": "导入失败",
                        "schema": {
                            "$ref": "#/definitions/models.APIResponse"
                        }
                    }
                }
            }
        },
        "/api/admin/questions/{id}": {
            "put": {
                "security": [
                    {
                        "BearerToken": []
                    }
                ],
                "description": "修改指定题目的内容、选项、答案等",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "题目管理"
                ],
                "summary": "更新题目",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "题目 ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "新题目信息",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.UpdateQuestionReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "更新成功",
                        "schema": {
                            "$ref": "#/definitions/models.APIResponse"
                        }
                    },
                    "400": {
                        "description": "参数错误",
                        "schema": {
                            "$ref": "#/definitions/models.APIResponse"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerToken": []
                    }
                ],
                "description": "删除指定题目",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "题目管理"
                ],
                "summary": "删除题目",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "题目 ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "删除成功",
                        "schema": {
                            "$ref": "#/definitions/models.APIResponse"
                        }
                    },
                    "400": {
                        "description": "无效的ID",
                        "schema": {
                            "$ref": "#/definitions/models.APIResponse"
                        }
                    },
                    "500": {
                        "description": "删除失败",
                        "schema": {
                            "$ref": "#/definitions/models.APIResponse"
                        }
                    }
                }
            }
        },
...
        "/api/admin/template": {
            "get": {
                "security": [
                    {
                        "BearerToken": []
                    }
                ],
                "description": "下载 Excel 格式的导入模板文件，含表头和格式说明",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/vnd.openxmlformats-officedocument.spreadsheetml.sheet"
                ],
                "tags": [
                    "题目管理"
                ],
                "summary": "下载考题导入模板",
                "responses": {
                    "200": {
                        "description": "Excel 模板文件",
                        "schema": {
                            "type": "file"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "models.APIResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "message": {
                    "type": "string"
                }
            }
        },
...
        "models.CreateQuestionReq": {
            "type": "object",
            "required": [
                "answer",
                "categoryId",
                "title",
                "type"
            ],
            "properties": {
                "answer": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "categoryId": {
                    "type": "integer"
                },
                "explanation": {
                    "type": "string"
                },
                "options": {
                    "description": "单选/多选必填；判断自动生成；填空/简答可为空",
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "title": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                }
            }
        },
...
        "models.PagedData": {
            "type": "object",
            "properties": {
                "list": {},
                "page": {
                    "type": "integer"
                },
                "pageSize": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                }
            }
        },
        "models.Question": {
            "type": "object",
            "properties": {
                "answer": {
                    "description": "JSON: [\"a\",\"b\"]（fill/essay 为纯文本答案）",
                    "type": "string"
                },
                "categoryId": {
                    "type": "integer"
                },
                "createdAt": {
                    "type": "string"
                },
                "explanation": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "options": {
                    "description": "JSON: {\"a\":\"选项A\",\"b\":\"选项B\"}（judge 自动为对/错；fill/essay 可为空）",
                    "type": "string"
                },
                "title": {
                    "type": "string"
                },
                "type": {
                    "description": "single / multiple / judge / fill / essay",
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
...
        "models.UpdateQuestionReq": {
            "type": "object",
            "properties": {
                "answer": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "categoryId": {
                    "type": "integer"
                },
                "explanation": {
                    "type": "string"
                },
                "options": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "title": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerToken": {
            "description": "输入 \"Bearer {token}\" 或通过 Cookie exam_token 传递",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    },
...
}